Move new modules interfaces in the js/modules package
