+++ js/common/bridge.go
@@ -280,72 +280,3 @@ func Bind(rt *goja.Runtime, v interface{}, ctxPtr *context.Context) map[string]i
 
 	return exports
 }
-
-// TODO move this
-// ModuleInstance is what a module needs to return
-type ModuleInstance interface {
-	ModuleInstanceCore
-	GetExports() Exports
-}
-
-func getInterfaceMethods() []string {
-	var t ModuleInstance
-	T := reflect.TypeOf(&t).Elem()
-	result := make([]string, T.NumMethod())
-
-	for i := range result {
-		result[i] = T.Method(i).Name
-	}
-
-	return result
-}
-
-// ModuleInstanceCore is something that will be provided to modules and they need to embed it in ModuleInstance
-type ModuleInstanceCore interface {
-	// we can add other methods here
-	// sealing field will help probably with pointing users that they just need to embed this in the
-	GetContext() context.Context
-}
-
-type Exports struct {
-	Default interface{}
-	Others  map[string]interface{}
-}
-
-func GenerateExports(v interface{}) Exports {
-	exports := make(map[string]interface{})
-	val := reflect.ValueOf(v)
-	typ := val.Type()
-	badNames := getInterfaceMethods()
-outer:
-	for i := 0; i < typ.NumMethod(); i++ {
-		meth := typ.Method(i)
-		for _, badname := range badNames {
-			if meth.Name == badname {
-				continue outer
-			}
-		}
-		name := MethodName(typ, meth)
-
-		fn := val.Method(i)
-		exports[name] = fn.Interface()
-	}
-
-	// If v is a pointer, we need to indirect it to access its fields.
-	if typ.Kind() == reflect.Ptr {
-		val = val.Elem()
-		typ = val.Type()
-	}
-	var mic ModuleInstanceCore // TODO move this out
-	for i := 0; i < typ.NumField(); i++ {
-		field := typ.Field(i)
-		if field.Type == reflect.TypeOf(&mic).Elem() {
-			continue
-		}
-		name := FieldName(typ, field)
-		if name != "" {
-			exports[name] = val.Field(i).Interface()
-		}
-	}
-	return Exports{Default: exports, Others: exports}
-}
