+++ js/modules/k6/metrics/metrics_test.go
@@ -31,6 +31,7 @@ import (
 	"gopkg.in/guregu/null.v3"
 
 	"go.k6.io/k6/js/common"
+	"go.k6.io/k6/js/modules"
 	"go.k6.io/k6/lib"
 	"go.k6.io/k6/lib/metrics"
 	"go.k6.io/k6/stats"
@@ -45,7 +46,7 @@ func (m *moduleInstanceImpl) GetContext() context.Context {
 	return *m.ctxPtr
 }
 
-func (m *moduleInstanceImpl) GetExports() common.Exports {
+func (m *moduleInstanceImpl) GetExports() modules.Exports {
 	panic("this needs to be implemented by the module")
 }
 
@@ -82,7 +83,7 @@ func TestMetrics(t *testing.T) {
 					*ctxPtr = common.WithInitEnv(*ctxPtr, &common.InitEnvironment{Registry: metrics.NewRegistry()})
 					m, ok := New().NewModuleInstance(&moduleInstanceImpl{ctxPtr: ctxPtr}).(*MetricsModule)
 					require.True(t, ok)
-					rt.Set("metrics", m.GetExports().Others) // This also should probably be done by some test package
+					rt.Set("metrics", m.GetExports().Named) // This also should probably be done by some test package
 					root, _ := lib.NewGroup("", nil)
 					child, _ := root.Group("child")
 					samples := make(chan stats.SampleContainer, 1000)
@@ -175,7 +176,7 @@ func TestMetricGetName(t *testing.T) {
 	ctx = common.WithInitEnv(ctx, &common.InitEnvironment{Registry: metrics.NewRegistry()})
 	m, ok := New().NewModuleInstance(&moduleInstanceImpl{ctxPtr: &ctx}).(*MetricsModule)
 	require.True(t, ok)
-	rt.Set("metrics", m.GetExports().Others) // This also should probably be done by some test package
+	rt.Set("metrics", m.GetExports().Named) // This also should probably be done by some test package
 	v, err := rt.RunString(`
 		var m = new metrics.Counter("my_metric")
 		m.name
@@ -200,7 +201,7 @@ func TestMetricDuplicates(t *testing.T) {
 	ctx = common.WithInitEnv(ctx, &common.InitEnvironment{Registry: metrics.NewRegistry()})
 	m, ok := New().NewModuleInstance(&moduleInstanceImpl{ctxPtr: &ctx}).(*MetricsModule)
 	require.True(t, ok)
-	rt.Set("metrics", m.GetExports().Others) // This also should probably be done by some test package
+	rt.Set("metrics", m.GetExports().Named) // This also should probably be done by some test package
 	_, err := rt.RunString(`
 		var m = new metrics.Counter("my_metric")
 	`)
