+++ js/modules/modules.go
@@ -21,21 +21,14 @@
 package modules
 
 import (
+	"context"
 	"fmt"
+	"reflect"
 	"strings"
 	"sync"
 
 	"go.k6.io/k6/js/common"
-	"go.k6.io/k6/js/modules/k6"
-	"go.k6.io/k6/js/modules/k6/crypto"
-	"go.k6.io/k6/js/modules/k6/crypto/x509"
-	"go.k6.io/k6/js/modules/k6/data"
-	"go.k6.io/k6/js/modules/k6/encoding"
-	"go.k6.io/k6/js/modules/k6/grpc"
-	"go.k6.io/k6/js/modules/k6/html"
 	"go.k6.io/k6/js/modules/k6/http"
-	"go.k6.io/k6/js/modules/k6/metrics"
-	"go.k6.io/k6/js/modules/k6/ws"
 )
 
 const extPrefix string = "k6/x/"
@@ -71,35 +64,97 @@ type HasModuleInstancePerVU interface {
 }
 
 // IsModuleV2 ... TODO better name
-type IsModuleV2 interface { // TODO rename?
-	NewModuleInstance(common.ModuleInstanceCore) common.ModuleInstance
+type IsModuleV2 interface {
+	NewModuleInstance(InstanceCore) ModuleInstance
 }
 
 // checks that modules implement HasModuleInstancePerVU
 // this is done here as otherwise there will be a loop if the module imports this package
 var _ HasModuleInstancePerVU = http.New()
 
-// GetJSModules returns a map of all js modules
+// GetJSModules returns a map of all registered js modules
 func GetJSModules() map[string]interface{} {
-	result := map[string]interface{}{
-		"k6":             k6.New(),
-		"k6/crypto":      crypto.New(),
-		"k6/crypto/x509": x509.New(),
-		"k6/data":        data.New(),
-		"k6/encoding":    encoding.New(),
-		"k6/net/grpc":    grpc.New(),
-		"k6/html":        html.New(),
-		"k6/http":        http.New(),
-		"k6/metrics":     metrics.New(),
-		"k6/ws":          ws.New(),
-	}
-
 	mx.Lock()
 	defer mx.Unlock()
+	result := make(map[string]interface{}, len(modules))
 
 	for name, module := range modules {
 		result[name] = module
 	}
 
 	return result
+}
+
+// ModuleInstance is what a module needs to return
+type ModuleInstance interface {
+	InstanceCore
+	GetExports() Exports
+}
+
+func getInterfaceMethods() []string {
+	var t ModuleInstance
+	T := reflect.TypeOf(&t).Elem()
+	result := make([]string, T.NumMethod())
+
+	for i := range result {
+		result[i] = T.Method(i).Name
+	}
+
+	return result
+}
+
+// InstanceCore is something that will be provided to modules and they need to embed it in ModuleInstance
+type InstanceCore interface {
+	// we can add other methods here
+	// sealing field will help probably with pointing users that they just need to embed this in the
+	GetContext() context.Context
+}
+
+// Exports is representation of ESM exports of a module
+type Exports struct {
+	// Default is what will be the `default` export of a module
+	Default interface{}
+	// Named is the named exports of a module
+	Named map[string]interface{}
+}
+
+// GenerateExports generates an Exports from a module akin to how common.Bind does now.
+// it also skips anything that is expected will not want to be exported such as methods and fields coming from
+// interfaces defined in this package.
+func GenerateExports(v interface{}) Exports {
+	exports := make(map[string]interface{})
+	val := reflect.ValueOf(v)
+	typ := val.Type()
+	badNames := getInterfaceMethods()
+outer:
+	for i := 0; i < typ.NumMethod(); i++ {
+		meth := typ.Method(i)
+		for _, badname := range badNames {
+			if meth.Name == badname {
+				continue outer
+			}
+		}
+		name := common.MethodName(typ, meth)
+
+		fn := val.Method(i)
+		exports[name] = fn.Interface()
+	}
+
+	// If v is a pointer, we need to indirect it to access its fields.
+	if typ.Kind() == reflect.Ptr {
+		val = val.Elem()
+		typ = val.Type()
+	}
+	var mic InstanceCore // TODO move this out
+	for i := 0; i < typ.NumField(); i++ {
+		field := typ.Field(i)
+		if field.Type == reflect.TypeOf(&mic).Elem() {
+			continue
+		}
+		name := common.FieldName(typ, field)
+		if name != "" {
+			exports[name] = val.Field(i).Interface()
+		}
+	}
+	return Exports{Default: exports, Named: exports}
 }
